Refactoring, clean-up and some bug fixes. Fixes #2 and #3.
This somewhat changes the way Biome IDs are handled.
The undefined/unpainted area biome IDs are now validated when the config
is being read, and if they are invalid, they will get reset to -1.
This also removes the need to always check if those IDs are valid
when they are being used, since they are validated when being set.
So they are always either -1, meaning use the biome from regular
terrain generation, or they hold a valid biome ID to use.

Also, the ColorToBiomeMapping map is now always used, and populated when
the config file is being read. The old useCustomColorMappings setting
is now gone/replaced by a new option which only controls how the mappings
are being created the first time new biomes that aren't in the config
yet are encountered. So even if the user wants to use simple
biomeID-to-color-channel style mapping, those colors now need to be set
in the ColorToBiomeMappings category for each biome.