Add template repeating functionality for the Single Template mode
There are two options: repeating the entire template image, or
repeating the biome of the edge-most pixel in the template image.

The repeating mode can be individually selected for each direction
(+x, -x, +z, -z).

If the two directions adjacent to the area beyond the corner of the
template image have the same repeating mode, then that area will use
that same setting. Ie. if both of those directions were set to use
the repeating template, then the area beyond the corner will also use
the repeating template. If those directions use the repeating edge mode,
then that area will use the biome from the corner pixel of the template image.

Note that a transparent corner pixel can be used here to get that
quadrant of the world to use either a set biome, or the default biome from
the regular terrain generator.